+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Next
+## 0.18.0 - 2024 Jun 7
 - Add arm64 support for macOS and Windows
 - Add View menu bar checkbox for album track grouping
 - Use album track grouping when sorting by album, date added, comments, genre, year & artist